# Exportable timer priority queue test suite
## Motivation

The timer priority queue is user-pluggable: the server runtime accepts any implementation of the `TimerPriorityQueue` interface, and Aiki ships two adapters — in-memory and Redis. The tests that define what the server expects from a queue lived as a single integration-test file inside the server package. It was bound to bun:test, and it got its queue from an internal harness helper that read `TIMER_PRIORITY_QUEUE_PROVIDER`, with a CI matrix supplying one run per provider.

That layout had two problems. An outside implementer had no way to run the tests — test files are not published, and a copy-paste goes stale the moment the expectations change. And the CI matrix reran the entire integration suite, database tests included, once per provider, just to flip which implementation that one file exercised.

## Solution

The suite body moves to `@aikirun/testing` as a function that takes its two environment-specific pieces as arguments. An implementer's whole test file looks like this:

```typescript
import { timerPriorityQueueTestSuite } from "@aikirun/testing/infra/timer";
import { describe, expect, test } from "vitest"; // or bun:test, or jest

timerPriorityQueueTestSuite({ describe, test, expect }, async (fn) => {
  const queue = createMyQueue(); // fresh and empty, once per test
  try {
    await fn(queue);
  } finally {
    // tear down
  }
});
```

The first argument is the caller's test framework — bun:test, vitest, and jest all satisfy the `{ describe, test, expect }` shape structurally, and the suite only uses portable matchers. The second is a per-test lifecycle combinator: hand the suite a fresh, empty queue, and clean up after it returns. Only the implementer knows how to build and reset their queue, so that part is theirs.

Each Aiki adapter now binds the suite in its own package, which makes Aiki's adapters the first consumers of the exported suite — the published tests and the enforced tests are the same code and cannot drift. The in-memory binder is a plain unit test needing no services. The Redis binder is an integration test connecting via `REDIS_URL`. The env variable, the harness helper, and the CI matrix are deleted; integration CI is back to one job, with Postgres and Redis services always up.

Passing the suite means the server's daemons behave correctly against the queue: pop ordering, due cutoffs, and waiter wake semantics including wake suppression. It deliberately does not certify durability, reconnects, or concurrent consumers — the queue is an at-least-once acceleration layer and the database remains the source of truth, so those stay the implementer's own testing job.

The database contract suite stays internal to the server package, selected by `DATABASE_PROVIDER` as before. The database is pluggable only by Aiki, so there is no outside implementer to serve.

`@aikirun/testing` is still a private workspace package, so the export becomes available to third parties when that package is published; the shape it ships with is the one in this PR.

One contributor-facing change: `bun run test:integration` now always needs a local Redis (previously only the redis matrix row did). CONTRIBUTING's integration-test section documents the full service setup.
