+++ .env.test.example
@@ -10,7 +10,4 @@ DATABASE_PROVIDER=pg
 DATABASE_URL=postgresql://user:password@localhost:5432/aiki_test
 DATABASE_MAX_CONNECTIONS=1
 
-# memory or redis
-TIMER_PRIORITY_QUEUE_PROVIDER=memory
-# Set if TIMER_PRIORITY_QUEUE_PROVIDER is redis
-# REDIS_URL=redis://localhost:6379
+REDIS_URL=redis://localhost:6379
