+++ sdk/server/package.json
@@ -45,11 +45,9 @@
 	"devDependencies": {
 		"@aikirun/lib": "workspace:*",
 		"@aikirun/memory": "workspace:*",
-		"@aikirun/redis": "workspace:*",
 		"@aikirun/testing": "workspace:*",
 		"drizzle-kit": "^0.31.8",
 		"fishery": "^2.4.0",
-		"ioredis": "^5.4.1",
 		"postgres": "^3.4.8"
 	},
 	"publishConfig": {
