+++ testing/package.json
@@ -7,6 +7,7 @@
 	"exports": {
 		"./client": "./src/client.ts",
 		"./infra/queue": "./src/infra/queue/index.ts",
+		"./infra/timer": "./src/infra/timer.ts",
 		"./data-factory/api/schedule": "./src/data-factory/api/schedule.ts",
 		"./data-factory/schedule": "./src/data-factory/schedule.ts",
 		"./data-factory/infra/queue": "./src/data-factory/infra/queue.ts",
